PLAT-245 -- Amend library as per PR.
