+++ test/endtoend/conftest.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+import pytest
+
+
+@pytest.fixture(scope="module")
+def db_uri():
+    return "mysql+mysqlconnector://root:@localhost/sqlalchemydiff"
